computer-use: renderer surfaces, bridge, and human-only lifecycle
Plugins → Computer Use owns the whole lifecycle, because none of it is
reachable by an agent: install a signed helper, grant it Accessibility
and Screen Recording, see which apps this session may drive, and remove
it along with every grant it held.

The approval card renders the payload for hazard actions — recipient,
text, destination — not just the app. "May use Mail" is not consent to
send this mail, and a card that only names the app is the failure G6
exists to prevent.

Adding the sidebar entry to the closed PluginNavEntry union made the
compiler name the three call sites that needed updating, which is what
that union is for.

Permission state refreshes on focus rather than on a timer: granting a
permission means leaving for System Settings and coming back, so focus
is exactly the moment the answer can have changed. Nothing polls.

CSS is component-scoped (app.css is closed to new rules), state is never
carried by colour alone, and the remove button says plainly that it
resets the system permissions too.

The specta command count is a hand-maintained tripwire; bumped 198 → 203
with the reason recorded beside it.

+++ apps/synth_desktop/src-tauri/src/computer_use/helper.rs
@@ -259,7 +259,7 @@ pub fn install(
 }
 
 /// What `remove` did, for the receipt. G7.
-#[derive(Clone, Debug, Serialize, Deserialize, PartialEq, Eq)]
+#[derive(Clone, Debug, Serialize, Deserialize, PartialEq, Eq, specta::Type)]
 #[serde(rename_all = "camelCase")]
 pub struct RemovalReport {
     pub bundle_removed: bool,
@@ -268,7 +268,9 @@ pub struct RemovalReport {
     /// Reset attempts macOS refused. Reported rather than swallowed: a grant
     /// left behind is exactly the uninstall residue G7 exists to catch.
     pub tcc_reset_failed: Vec<String>,
-    pub allowlist_entries_removed: usize,
+    /// u32 rather than usize: specta forbids BigInt-style types across the
+    /// bridge, and nobody has more than four billion allowlist entries.
+    pub allowlist_entries_removed: u32,
 }
 
 /// TCC service names, as `tccutil` spells them.
